+++ server/main/src/constant.rs
@@ -74,7 +74,7 @@ lazy_static! {
         String::from("glsl"),
     ]);
     pub static ref RE_MACRO_INCLUDE: Regex = Regex::new(r#"^\s*(?:#include)\s+"(.+)""#).unwrap();
-    pub static ref RE_MACRO_VERSION: Regex = Regex::new(r#"^\s*(?:#version)"#).unwrap();
+    pub static ref RE_MACRO_VERSION: Regex = Regex::new(r#"(?m)^\s*(?:#version).*$"#).unwrap();
     pub static ref RE_MACRO_LINE: Regex = Regex::new(r#"^\s*(?:#line)"#).unwrap();
 }