+++ server/main/src/file/temp_file.rs
@@ -3,8 +3,6 @@ use std::{cell::RefCell, collections::HashMap, fs::read_to_string, path::PathBuf
 use logging::warn;
 use tree_sitter::{Parser, Tree};
 
-use crate::constant::{OPTIFINE_MACROS, RE_MACRO_INCLUDE, RE_MACRO_LINE, RE_MACRO_VERSION};
-
 use super::*;
 
 impl TempFile {
@@ -66,9 +64,6 @@ impl TempFile {
         let mut file_id = 0;
         let file_name = file_path.display().to_string();
 
-        // If we are in the debug folder, do not add Optifine's macros
-        let mut macro_inserted = self.pack_path.parent().unwrap().file_name().unwrap() == "debug";
-
         self.content.borrow().lines().enumerate().for_each(|line| {
             if let Some(capture) = RE_MACRO_INCLUDE.captures(line.1) {
                 let path = capture.get(1).unwrap().as_str();
@@ -90,14 +85,22 @@ impl TempFile {
             } else {
                 temp_content += line.1;
                 temp_content += "\n";
-                // If we are not in the debug folder, add Optifine's macros for correct linting
-                if !macro_inserted && RE_MACRO_VERSION.is_match(line.1) {
-                    temp_content += OPTIFINE_MACROS;
-                    temp_content += &format!("#line {} 0\t//{}\n", line.0 + 2, file_name);
-                    macro_inserted = true;
-                }
             }
         });
+
+        // Move #version to the top line
+        if let Some(capture) = RE_MACRO_VERSION.captures(&temp_content) {
+            let version = capture.get(0).unwrap();
+            let mut version_content = version.as_str().to_owned() + "\n";
+
+            temp_content.replace_range(version.start()..version.end(), "");
+            // If we are not in the debug folder, add Optifine's macros
+            if self.pack_path.parent().unwrap().file_name().unwrap() != "debug" {
+                version_content += OPTIFINE_MACROS;
+            }
+            version_content += &format!("#line 1 0\t//{}\n", file_name);
+            temp_content.insert_str(0, &version_content);
+        }
 
         Some((self.file_type, temp_content))
     }